% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_circle.R
\name{area_circle}
\alias{area_circle}
\title{Calculate the area of a circle.}
\usage{
area_circle(r)
}
\arguments{
\item{r}{radius of a circle.}
}
\value{
The area of the circle.
}
\description{
Calculate the area of a circle.
}
\examples{
area_circle(2)
area_circle(pi)
}
