% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_ellipse.R
\name{area_ellipse}
\alias{area_ellipse}
\title{Calculate the area of an ellipse}
\usage{
area_ellipse(r1, r2)
}
\arguments{
\item{r1}{semi-major radius of an ellipse.}

\item{r2}{semi-minor radius of an ellipse.}
}
\value{
The area of an ellipse.
}
\description{
Calculate the area of an ellipse
}
\examples{
area_ellipse(0.5, 0.25)
area_ellipse(2, 2)
}
