% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_parallelogram.R
\name{area_parallelogram}
\alias{area_parallelogram}
\title{Calculate the area of a parallelogram}
\usage{
area_parallelogram(b, h)
}
\arguments{
\item{b}{base of the parallelogram.}

\item{h}{height of the parallelogram.}
}
\value{
The area of the parallelogram
}
\description{
Calculate the area of a parallelogram
}
\examples{
area_parallelogram(4, 8)
area_parallelogram(16, 2.5)
}
