% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_rectangle.R
\name{area_rectangle}
\alias{area_rectangle}
\title{Calculate the area of a rectangle}
\usage{
area_rectangle(l, w)
}
\arguments{
\item{l}{length of the rectangle.}

\item{w}{width of the rectangle.}
}
\value{
The area of the rectangle.
}
\description{
Calculate the area of a rectangle
}
\examples{
area_rectangle(4,6)
area_rectangle(10,20)
}
