% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_square.R
\name{area_square}
\alias{area_square}
\title{Calculate the area of a square}
\usage{
area_square(s)
}
\arguments{
\item{s}{length of a side.}
}
\value{
The area of the square
}
\description{
Calculate the area of a square
}
\examples{
area_square(2)
area_square(10)
}
