% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_trapezoid.R
\name{area_trapezoid}
\alias{area_trapezoid}
\title{Calculate the area of a trapezoid.}
\usage{
area_trapezoid(b1, b2, h)
}
\arguments{
\item{b1}{bottom base of a trapezoid.}

\item{b2}{top base of a trapezoid.}

\item{h}{height of a trapezoid.}
}
\value{
The area of the trapezoid.
}
\description{
Calculate the area of a trapezoid.
}
\examples{
area_trapezoid(3, 5, 16)
area_trapezoid(0.5, 0.5, 2)
}
