% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_triangle.R
\name{area_triangle}
\alias{area_triangle}
\title{Calculate the area of a triangle.}
\usage{
area_triangle(b, h)
}
\arguments{
\item{b}{base of a triangle.}

\item{h}{height of a triangle.}
}
\value{
The area of the triangle.
}
\description{
Calculate the area of a triangle.
}
\examples{
area_triangle(2, 6)
area_triangle(3.5, 4)
}
